// src/types/device.ts
export type DeviceMetrics = {
  deviceHeight: number;
  deviceLeft: number;
  deviceTop: number;
  deviceWidth: number;
  expandedScale: number;
  holeCenterX: number;
  holeCenterY: number;
  holeLeft: number;
  holeRadius: number;
  holeTop: number;
  screenHeight: number;
  screenWidth: number;
  zoomOriginX: number;
  zoomOriginY: number;
};